move test and endpoint
